--- src/main/kotlin/jp/trap/plutus/pteron/features/project/domain/ProjectTransaction.kt
@@ -1,5 +1,0 @@
-package jp.trap.plutus.pteron.features.project.domain
-
-interface ProjectTransaction {
-    suspend fun <T> runInTransaction(block: () -> T): T
-}
