+++ src/main/kotlin/jp/trap/plutus/pteron/features/project/domain/transaction/ProjectTransaction.kt
@@ -1,0 +1,5 @@
+package jp.trap.plutus.pteron.features.project.domain.transaction
+
+interface ProjectTransaction {
+    suspend fun <T> runInTransaction(block: suspend () -> T): T
+}
